[protocol] split server and client side protocol in their own files.
Client and server side use different dependencies (aioh2 is only used
client side for instance). There is not much in common between those 2,
so let's split them.
